Int16 -> Int32

+++ src/basins/mapping/attractor_mapping.jl
@@ -142,7 +142,7 @@ the partitioning happens directly on the hyperplane the Poincaré map operates o
 corresponding to the state space partitioning indicated by `grid`.
 """
 function basins_of_attraction(mapper::AttractorMapper, grid::Tuple; kwargs...)
-    basins = zeros(Int16, map(length, grid))
+    basins = zeros(Int32, map(length, grid))
     I = CartesianIndices(basins)
     A = Dataset([generate_ic_on_grid(grid, i) for i in I])
     fs, labels, attractors = basins_fractions(mapper, A; kwargs...)
